fix(tou): add role-scoped agreement pages
