+++ src/app/[locale]/agreement/components/ReacceptButton/index.tsx
@@ -2,17 +2,22 @@ import { Button, useTheme } from "@mui/material"
 import { useRouter } from "next/navigation"
 import { useTranslation } from "react-i18next"
 
+import type { ServiceAgreementPartyInput } from "@/app/api/service-agreement/interface"
 import { useAcceptToU } from "@/hooks/useToUReacceptance"
 
 /// Party-aware re-acceptance CTA for accounts that signed an older ToU version
 /// or declined the current one. Same flow as the re-acceptance modal (borrower
 /// accounts sign with their organization name); the legacy SignButton stays
 /// lender-only for first-time onboarding and its old-table dual-write.
-export const ReacceptButton = () => {
+export const ReacceptButton = ({
+  party,
+}: {
+  party: ServiceAgreementPartyInput
+}) => {
   const theme = useTheme()
   const { t } = useTranslation()
   const router = useRouter()
-  const accept = useAcceptToU()
+  const accept = useAcceptToU(party)
 
   const handleSign = () => {
     // Return to where the user came from (re-acceptance modal, create-market
