+++ src/app/[locale]/borrower/agreement/page.tsx
@@ -1,0 +1,5 @@
+import { AgreementPage } from "@/app/[locale]/agreement/components/AgreementPage"
+
+export default function BorrowerAgreement() {
+  return <AgreementPage party="Borrower" />
+}
