+++ src/app/[locale]/borrower/create-market/page.tsx
@@ -49,7 +49,6 @@ import {
   saveCreateMarketSigningDraft,
 } from "@/store/slices/createMarketSigningDraftsSlice/createMarketSigningDraftsSlice"
 import { removePendingSafeMessage } from "@/store/slices/pendingSafeMessagesSlice/pendingSafeMessagesSlice"
-import { setTouModalOpen } from "@/store/slices/touModalSlice/touModalSlice"
 import { COLORS } from "@/theme/colors"
 
 import { BasicSetupForm } from "./components/Forms/BasicSetupForn"
@@ -710,7 +709,7 @@ export default function CreateMarketPage() {
           <Button
             variant="contained"
             size="large"
-            onClick={() => dispatch(setTouModalOpen(true))}
+            onClick={() => router.push(ROUTES.borrower.agreement)}
             sx={{ minWidth: "220px" }}
           >
             Review Terms of Use
