+++ src/app/[locale]/lender/agreement/page.tsx
@@ -1,0 +1,5 @@
+import { AgreementPage } from "@/app/[locale]/agreement/components/AgreementPage"
+
+export default function LenderAgreement() {
+  return <AgreementPage party="Lender" />
+}
